+++ app/controllers/tags_controller.rb
@@ -3,7 +3,7 @@ class TagsController < ApplicationController
 
   def index
     @tag_scope = params[:scope].to_s == "story" ? "story" : "universe"
-    allowed_types = @tag_scope == "story" ? %w[section] : %w[character relation location event item ownership]
+    allowed_types = @tag_scope == "story" ? %w[section scene] : %w[character relation location event item ownership]
     @tag_type = params[:taxonomy].to_s
     @tag_type = allowed_types.include?(@tag_type) ? @tag_type : allowed_types.first
   end