+++ app/services/development/universe_data_registry.rb
@@ -29,6 +29,7 @@ module Development
       ModelDefinition.new(model_name: "UniverseMembership", file_name: "universe_memberships", scope: :universe),
       ModelDefinition.new(model_name: "Story", file_name: "stories", scope: :universe),
       ModelDefinition.new(model_name: "SectionTag", file_name: "section_tags", scope: :story, positioned: true),
+      ModelDefinition.new(model_name: "SceneTag", file_name: "scene_tags", scope: :story, positioned: true),
       ModelDefinition.new(model_name: "Section", file_name: "sections", scope: :story, positioned: true),
       ModelDefinition.new(model_name: "LocationTag", file_name: "location_tags", scope: :universe, positioned: true),
       ModelDefinition.new(model_name: "Location", file_name: "locations", scope: :universe, positioned: true),